Addressing reviewer requests.
